refactor: restructure board-related files to follow FSD
